+++ source/Obstacles/CarlingFish.cpp
@@ -14,7 +14,7 @@ CarlingFish::CarlingFish(SimulationData&s, ArgumentParser&p, Real C[2])
   : Fish(s,p,C) {
   const Real ampFac = p("-amplitudeFactor").asDouble(1.0);
   myFish = new AmplitudeFish(length, Tperiod, phaseShift, sim.minH, ampFac);
-  if( s.verbose ) printf("[CUP2D] - AmplitudeFish %d %f %f %f\n",myFish->Nm, length, Tperiod, phaseShift);
+  if( s.verbose ) printf("[CUP2D] - AmplitudeFish %d %f %f %f\n",myFish->Nm, (double)length, (double)Tperiod, (double)phaseShift);
 }
 
 void CarlingFish::create(const std::vector<BlockInfo>& vInfo) {
