+++ source/Obstacles/NeuroKinematicFish.cpp
@@ -151,8 +151,8 @@ public:
         undulatoryCurvatureScheduler.transition(t_current, t_current, this->t_next, undulatoryCurvatureValues, useCurrentDerivative);
         tauTailScheduler.transition(t_current, t_current, this->t_next, tailPhase, useCurrentDerivative);
 
-        printf("Performing a burst with timingFactor %f, and modulationFactor %f\n", timingFactor, modulationFactor);
-        printf("t_next is: %f\n", this->t_next);
+        printf("Performing a burst with timingFactor %f, and modulationFactor %f\n", (double)timingFactor, (double)modulationFactor);
+        printf("t_next is: %f\n", (double)this->t_next);
     }
 
     void scoot(const Real t_current, const std::vector<Real> &a) {
@@ -193,8 +193,8 @@ public:
         undulatoryCurvatureScheduler.transition(t_current, t_current, this->t_next, undulatoryCurvatureValues, useCurrentDerivative);
         tauTailScheduler.transition(t_current, t_current, this->t_next, tailPhase, useCurrentDerivative);
 
-        printf("Performing a scoot with timingFactor %f, and modulationFactor %f\n", timingFactor, modulationFactor);
-        printf("t_next is: %f\n", this->t_next);
+        printf("Performing a scoot with timingFactor %f, and modulationFactor %f\n", (double)timingFactor, (double)modulationFactor);
+        printf("t_next is: %f\n", (double)this->t_next);
     }
 
     void coast(const Real t_current, const std::vector<Real> &a) {
@@ -235,8 +235,8 @@ public:
     undulatoryCurvatureScheduler.transition(t_current, t_current, this->t_next, undulatoryCurvatureValues, useCurrentDerivative);
     tauTailScheduler.transition(t_current, t_current, this->t_next, tailPhase, useCurrentDerivative);
 
-    printf("Performing a coast with timingFactor %f, and modulationFactor %f\n", timingFactor, modulationFactor);
-    printf("t_next is: %f\n", this->t_next);
+    printf("Performing a coast with timingFactor %f, and modulationFactor %f\n", (double)timingFactor, (double)modulationFactor);
+    printf("t_next is: %f\n", (double)this->t_next);
 }
 
     void hybrid(const Real t_current, const std::vector<Real> &a)
@@ -285,8 +285,8 @@ public:
         baselineCurvatureScheduler.transition(t_current, t_current, this->t_next, baselineCurvatureValues, useCurrentDerivative);
         undulatoryCurvatureScheduler.transition(t_current, t_current, this->t_next, undulatoryCurvatureValues, useCurrentDerivative);
         tauTailScheduler.transition(t_current, t_current, this->t_next, tailPhase, useCurrentDerivative);
-        printf("Performing a hybrid action with beta %f, kappa %f, c %f\n", lastBeta, lastKappa, lastC);
-        printf("t_next is: %f\n", this->t_next);
+        printf("Performing a hybrid action with beta %f, kappa %f, c %f\n", (double)lastBeta, (double)lastKappa, (double)lastC);
+        printf("t_next is: %f\n", (double)this->t_next);
     }
 
     void spike(const Real t_current, const std::vector<Real> &a) {
@@ -304,8 +304,8 @@ public:
         this->t_next = t_current + lastFireTime;
 
         neuroKinematicScheduler.Spike(t_current, lastAmplitude, lastDelay, lastFireTime);
-        printf("Spiking at %f, with amplitude %f, delay %f, fire-time %f\n", t_current, lastAmplitude, lastDelay, lastFireTime);
-        printf("t_next is: %f\n", this->t_next);
+        printf("Spiking at %f, with amplitude %f, delay %f, fire-time %f\n", (double)t_current, (double)lastAmplitude, (double)lastDelay, (double)lastFireTime);
+        printf("t_next is: %f\n", (double)this->t_next);
 
     }
 };
@@ -386,7 +386,7 @@ NeuroKinematicFish::NeuroKinematicFish(SimulationData&s, ArgumentParser&p, Real 
 {
     const Real ampFac = p("-amplitudeFactor").asDouble(1.0);
     myFish = new NeuroFish(length, Tperiod, phaseShift, sim.minH, ampFac);
-    if( s.verbose ) printf("[CUP2D] - NeuroFish %d %f %f %f\n",myFish->Nm, length, Tperiod, phaseShift);
+    if( s.verbose ) printf("[CUP2D] - NeuroFish %d %f %f %f\n",myFish->Nm, (double)length, (double)Tperiod, (double)phaseShift);
 }
 
 void NeuroKinematicFish::resetAll() {
